feat(predicate): added `BoosterPredicates.withParameters()`
Added a booster predicate that uses a multi-valued map with parameters
as used generally by HTTP/REST requests. Each key in the map will be used
as a "property path" selector for the boosters to select. This means that
"id" will match a booster's "id" field while "mission.name" will match the
"name" field of the booster's `Mission` object.
The map value that's associated with the key will be compared against
the string value of the selected booster property and if they are the
same the booster will be included in the result.
If no value was given for the key (ie the value was an empty string) then
"true" will be assumed. If a key selects a non-existing property then its
value is taken to be "false". (This means that selecting `nonExistingProperty`
will return no results but `nonExistingProperty=false` will return all results)

See fabric8-launcher/launcher-planning#143